Fix misplaced methods
